requirements.txt
flask
flask-bootstrap4
paho-mqtt
python-dotenv
